ファイルのダウンロードアクション（TODO：湖底になっている）の場合、トップページに戻す

+++ Error/NetCommonsExceptionRenderer.php
@@ -69,7 +69,8 @@ class NetCommonsExceptionRenderer extends ExceptionRenderer {
 		if ($message === 'The request has been black-holed') {
 			$message = __d('net_commons', 'The request has been black-holed');
 			$redirect = $this->controller->request->referer(true);
-		} elseif (get_class($error) === 'ForbiddenException') {
+		} elseif (get_class($error) === 'ForbiddenException' &&
+				! in_array($this->controller->request->params['action'], ['download'], true)) {
 			$redirect = Router::url('/auth/login');
 			$this->controller->Session->destroy();
 			if (! $this->controller->request->is('ajax')) {
